Split basic_result from webview.h
